src/data_collection/batch_requirements.txt
google-genai
requests
aiohttp
tqdm
Pillow
matplotlib
